fix: resolve all type-checker diagnostics and latent runtime bugs
- Models: TYPE_CHECKING imports for forward-referenced relationship types;
  all Numeric columns typed as Mapped[Decimal] (matches SQLAlchemy runtime)
- products.py: add missing InventoryMovement import (unreachable /inventory
  endpoint would raise NameError); fix float*Decimal valuation math
- customers/suppliers statements: replace .model_validate() calls on ORM
  model classes (AttributeError at runtime) with proper Pydantic response
  schemas; invoice rows serialized via table columns
- reports.py: narrow optional parsed dates before service calls; restructure
  invoice PDF party lookup to avoid cross-type attribute access; handle
  Optional workbook active sheet
- accounting.py: narrow optional parsed dates before get_profit_loss
- sale_return.py: guard against missing invoice line item
- schemas/common.py: single-use TypeVar removed from ok(); tag.py: TagUpdate
  no longer overrides base field incompatibly
- services/reports.py: drop invalid sqlalchemy.orm.exc import

Verified: pyright 0 errors/0 warnings; pytest 51 passed 1 skipped;
svelte-check 0 errors/0 warnings; frontend build OK

+++ backend/app/models/bank_account.py
@@ -1,4 +1,6 @@
 from datetime import datetime
+from decimal import Decimal
+
 from sqlalchemy import Boolean, DateTime, Numeric, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -14,8 +16,8 @@ class BankAccount(TimestampMixin, Base):
     iban: Mapped[str | None] = mapped_column(String(34), nullable=True)
     account_number: Mapped[str | None] = mapped_column(String(50), nullable=True)
     bank_name: Mapped[str | None] = mapped_column(String(255), nullable=True)
-    opening_balance: Mapped[float] = mapped_column(Numeric(15, 2), nullable=False, default=0)
-    current_balance: Mapped[float] = mapped_column(Numeric(15, 2), nullable=False, default=0)
+    opening_balance: Mapped[Decimal] = mapped_column(Numeric(15, 2), nullable=False, default=0)
+    current_balance: Mapped[Decimal] = mapped_column(Numeric(15, 2), nullable=False, default=0)
     active: Mapped[bool] = mapped_column(Boolean, default=True, nullable=False, index=True)
     notes: Mapped[str | None] = mapped_column(Text, nullable=True)
     deleted_at: Mapped[datetime | None] = mapped_column(DateTime(timezone=True), nullable=True, index=True)
